Move logger from internal to pkg
