+++ cmd/dev/app/testserver/testserver.go
@@ -25,11 +25,11 @@ import (
 	serverconfig "github.com/mindersec/minder/internal/config/server"
 	"github.com/mindersec/minder/internal/controlplane/metrics"
 	"github.com/mindersec/minder/internal/db/embedded"
-	"github.com/mindersec/minder/internal/logger"
 	"github.com/mindersec/minder/internal/metrics/meters"
 	"github.com/mindersec/minder/internal/providers/ratecache"
 	provtelemetry "github.com/mindersec/minder/internal/providers/telemetry"
 	"github.com/mindersec/minder/internal/service"
+	"github.com/mindersec/minder/pkg/logger"
 )
 
 // CmdTestServer starts a test server for integration testing.
