+++ cmd/server/app/encryption_rotate.go
@@ -19,7 +19,7 @@ import (
 	serverconfig "github.com/mindersec/minder/internal/config/server"
 	"github.com/mindersec/minder/internal/crypto"
 	"github.com/mindersec/minder/internal/db"
-	"github.com/mindersec/minder/internal/logger"
+	"github.com/mindersec/minder/pkg/logger"
 )
 
 // number of secrets to re-encrypt per batch
