+++ cmd/server/app/migrate_up.go
@@ -20,7 +20,7 @@ import (
 	"github.com/mindersec/minder/internal/authz"
 	"github.com/mindersec/minder/internal/config"
 	serverconfig "github.com/mindersec/minder/internal/config/server"
-	"github.com/mindersec/minder/internal/logger"
+	"github.com/mindersec/minder/pkg/logger"
 )
 
 // upCmd represents the up command
