+++ cmd/server/app/migrate_version.go
@@ -17,7 +17,7 @@ import (
 	"github.com/mindersec/minder/database"
 	"github.com/mindersec/minder/internal/config"
 	serverconfig "github.com/mindersec/minder/internal/config/server"
-	"github.com/mindersec/minder/internal/logger"
+	"github.com/mindersec/minder/pkg/logger"
 )
 
 // versionCmd represents the version command
