+++ cmd/server/app/serve.go
@@ -23,11 +23,11 @@ import (
 	serverconfig "github.com/mindersec/minder/internal/config/server"
 	cpmetrics "github.com/mindersec/minder/internal/controlplane/metrics"
 	"github.com/mindersec/minder/internal/db"
-	"github.com/mindersec/minder/internal/logger"
 	"github.com/mindersec/minder/internal/metrics/meters"
 	"github.com/mindersec/minder/internal/providers/ratecache"
 	provtelemetry "github.com/mindersec/minder/internal/providers/telemetry"
 	"github.com/mindersec/minder/internal/service"
+	"github.com/mindersec/minder/pkg/logger"
 )
 
 var serveCmd = &cobra.Command{
