+++ internal/controlplane/handlers_token.go
@@ -20,9 +20,9 @@ import (
 	"google.golang.org/protobuf/reflect/protoregistry"
 
 	"github.com/mindersec/minder/internal/auth/jwt"
-	"github.com/mindersec/minder/internal/logger"
 	"github.com/mindersec/minder/internal/util"
 	minder "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
+	"github.com/mindersec/minder/pkg/logger"
 )
 
 // TokenValidationInterceptor is a server interceptor that validates the bearer token
