+++ internal/controlplane/handlers_user.go
@@ -25,10 +25,10 @@ import (
 	"github.com/mindersec/minder/internal/db"
 	"github.com/mindersec/minder/internal/flags"
 	"github.com/mindersec/minder/internal/invites"
-	"github.com/mindersec/minder/internal/logger"
 	"github.com/mindersec/minder/internal/projects"
 	"github.com/mindersec/minder/internal/util"
 	pb "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
+	"github.com/mindersec/minder/pkg/logger"
 )
 
 // CreateUser is a service for user self registration
