+++ internal/reconcilers/entity_add.go
@@ -13,9 +13,9 @@ import (
 	"github.com/rs/zerolog"
 
 	"github.com/mindersec/minder/internal/entities/properties"
-	"github.com/mindersec/minder/internal/logger"
 	"github.com/mindersec/minder/internal/reconcilers/messages"
 	pb "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
+	"github.com/mindersec/minder/pkg/logger"
 )
 
 // handleEntityAddEvent handles the entity add event.
