+++ pkg/logger/telemetry_store_watermill_test.go
@@ -16,9 +16,9 @@ import (
 	serverconfig "github.com/mindersec/minder/internal/config/server"
 	"github.com/mindersec/minder/internal/engine/entities"
 	"github.com/mindersec/minder/internal/events"
-	"github.com/mindersec/minder/internal/logger"
 	"github.com/mindersec/minder/internal/util/testqueue"
 	minderv1 "github.com/mindersec/minder/pkg/api/protobuf/go/minder/v1"
+	"github.com/mindersec/minder/pkg/logger"
 )
 
 func TestTelemetryStoreWMMiddlewareLogsRepositoryInfo(t *testing.T) {
